cashier-system/cashier-system/requirements.txt
mysql-connector-python
openpyxl
pandas
tkinter